+++ CHANGELOG.md
@@ -5,9 +5,21 @@ Please prefix each entry with one of:
 - [fix]
 - [documentation]
 
+
 # Next Release
+- [feature] Support for uploading to artifactory with the Teamscale default artifact storage schema.
+
+# v22.8.0
+- [feature] support for .NET core PDBs
+
+# v22.7.0
+- [feature] Upload Daemon sends distinct user agent header for Teamscale uploads
+- [fix] Updated Newtonsoft.JSON to version 13.0.1
+
+
+# v22.4.0
 - [feature] Caching symbols for better performance when converting many trace files to line coverage.
+- [fix] Upload Daemon could be started multiple times
 
 # v19.8.0
 - [fix] async upload bug
-